refactor: change definition file to normal file
