+++ src/background/message-distributor.ts
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { BrowserAdapter } from '../common/browser-adapters/browser-adapter';
+import { Tab } from '../common/itab';
 import { Logger } from '../common/logging/logger';
 import { InterpreterMessage } from '../common/message';
-import { Tab } from './../common/itab.d';
 import { GlobalContext } from './global-context';
 import { TabToContextMap } from './tab-context';
